fix: api_userdata inside api_metadata
